refactor: move split_chunks from render.rs to layout/split.rs
Resolves circular dependency: render.rs imported from layout, layout
imported split_chunks from render. Now both directions flow from layout.
